Safari-611.1.10/LayoutTests/accessibility/gtk/text-at-offset-preformatted-expected.txt

	
	first line
	second line
	
This tests the ability to get element text for different granularities and offsets.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS content.lineAtOffset(0) is '\t\n, 0, 2'
PASS content.lineAtOffset(2) is '\tfirst line\n, 2, 14'
PASS content.lineAtOffset(14) is '\tsecond line\n, 14, 27'
PASS successfullyParsed is true

TEST COMPLETE
